Fix: handle lane splitting
Problem: Lane-follower can't handle when the current lane splits into
two lanes. Because it takes the component centroid closest to the
previous iteration's, when the lane splits up, it can't accurately
decide which white line is the edge of the lane it's currently
following.

Solution: Using bit masking, erase the right line, so the vehicle only
sees the left lane and the dashed line, effectively making the left lane
the only lane.

Pitfall: The region used for bit-masking is hard-coded for this specific
road, may not work as a general solution.