Fix: jerking from left to right
Problem: Since one of the line the vehicle was following was a dashed
line, it had problem applying lane-following logic for the places where
the dashed line doesn't exist. The result is the vehicle advancing
forward with a left-right jerking.

Solution: Use the leftmost and rightmost line, but keep the vehicle in
the left lane, i.e., left of the dashed line.